Issue #4575: Modified JavaNCSSCheckTest.java and moved its input files to the javancss subdirectory
